Dev/init fix (#1)
* update# Remove expired IsReadyProcessAttestation

* fix# BlockStatus->Status adjustment type is uint8

* fix# Fix code Attestation->RecoverSigner->signer

* fix# Advance the logic to judge whether it already exists

* fix# Optimize code

* fix# Optimization log

* fix# Send attestations that the relevant methods have been replaced asynchronously

* update# Optimize block query

+++ internal/ethapi/api.go
@@ -856,7 +856,7 @@ func (s *PublicBlockChainAPI) getDoubleSignPunishTransactions(block *types.Block
 	return transactions, nil
 }
 
-func (s *PublicBlockChainAPI) GetBlockPredictStatus(ctx context.Context, hash common.Hash, number rpc.BlockNumber) (uint64, error) {
+func (s *PublicBlockChainAPI) GetBlockPredictStatus(ctx context.Context, hash common.Hash, number rpc.BlockNumber) (uint8, error) {
 	block, err := s.b.BlockByHash(ctx, hash)
 	if err != nil || block == nil {
 		return 0, err
