+++ internal/ethapi/backend.go
@@ -63,7 +63,7 @@ type Backend interface {
 	CurrentBlock() *types.Block
 	BlockByNumber(ctx context.Context, number rpc.BlockNumber) (*types.Block, error)
 	BlockByHash(ctx context.Context, hash common.Hash) (*types.Block, error)
-	BlockPredictStatus(ctx context.Context, hash common.Hash, number rpc.BlockNumber) (uint64, error)
+	BlockPredictStatus(ctx context.Context, hash common.Hash, number rpc.BlockNumber) (uint8, error)
 	BlockByNumberOrHash(ctx context.Context, blockNrOrHash rpc.BlockNumberOrHash) (*types.Block, error)
 	StateAndHeaderByNumber(ctx context.Context, number rpc.BlockNumber) (*state.StateDB, *types.Header, error)
 	StateAndHeaderByNumberOrHash(ctx context.Context, blockNrOrHash rpc.BlockNumberOrHash) (*state.StateDB, *types.Header, error)
